updated index.html  and scripts.js
added a div element and button element for the modal of the pokemon, in index.html.

updating the showDetails() to show a modal upon clicking on each pokemon:

Created a div element for the modal container, on the document node, then appended them to the unordered list in the html file.

also removed unwanted comments.

Created a div element for the modal and then appended to a modal container, previously stated.

Created sub- elements of the modal on document, then appended them to the modal.

also added parameters to the showDetails function to better reference the pokemon's  name and height.