Update scripts.js
corrected a mistake with the name of the showModal() function.

updated the parameters of the showModal() function.

add an attribute to the img element in the modal.

re-added showDetails() function into my code, which was previously overwritten:  once the loadDetails() is successful, it will load the showModal().

added loadDetails and showDetails into the return object of the pokemonRepository.

updated indentation.