updated styles.css
Deleted  code for the modal's styling- With added Bootstrap UI modal  component, there is no need to styling the modal.

kept the styling of the 'is-visible' class, to show the container upon a 'click' event.

removed styling for the 'tittle' and 'text' class- which were arguments for the previous dynamically created modal code of the showModal() function.

changed colour of the background and button inside the Pokemon list colour on line 2 and 32 respectively.