updated scripts.js
updated the selection of the '#modalContainer'.

incorporated Bootstrap's css classes onto the dynamically created elements in addListItem()- giving the <button > a data-toggle attribute to open Bootstrap's UI modal component.

Removed an older version of the showModal() which didn't incorporated Bootstrap's UI modal component.

added an extra object property, created dynamically in loadDetails(), to access the pokemon's type(s).

used jQuery syntax to create newly added the show Modal() function.

added a general function to hide the modal- using this as a reference to the three event handlers to close the modal.

removed old comments/ commented code, throughout the file.

updated the argument of the showModal() on line 157.