updated dist folder and index.html
created a new file of the minified the style.css file in the dist/ folder.

In index.html, changed the stylesheet of the .css file to the .css in the dist/ folder- to increase network performance.

changed the src of the <script> on the .js files, to the minified .js files- to increease network performance.